tauri: hardening: disable all commands by default
[The Tauri's API example app](https://github.com/tauri-apps/tauri/blob/cde0ff7798a46712f69b34aab952209f45500fe9/examples/api/src-tauri/build.rs#L9-L21)
was used as a reference for the implementation of this commit.

This https://github.com/tauri-apps/tauri/pull/9008 says:
> By default all app commands are still allowed,
> so if you do not use `tauri_build::Attributes::app_manifest`
> you're not affected by this change.

Tauri docs don't seem to say much about how to disable this
"allow all commands by default" thing.

I have tested this change and verified that commands
are not available in the HTML email viewer contents webview,
and that only the required ones are available in its header webview.

Note that the upcoming webxdc MR
(https://github.com/deltachat/deltachat-desktop/pull/4740)
will need to be adjusted
to specify the allowed commands for the webxdc app windows.

Co-authored-by: SimonLaux <mobile.info@simonlaux.de>